Replace 'dotenv' for Defines.inc
Most of these directives provide conditional compilation for developer-only features. Storing them in memory poses security issues. Even though in this case (Sentry) it does end up in the memory, it is only for convenience and not used to add any critical code, we should not promote indirect usage of directives.